# use 32 bit index to improve activation ops